// src/ts/types/gprs.types.ts
export type GPRSResponseType = {
  attributes: {
    activeEnergy: boolean;
    amperageLimit: number;
    apnName: string;
    apnPassword: string;
    apnUsername: string;
    closeDay: number;
    closeHour: number;
    domainIp: string;
    impulseNumber: number;
    instantPower: boolean;
    localIp: string;
    maxPower: boolean;
    networkSettings: boolean;
    pagingDelay: number;
    payloadLimit: number;
    pingIp: string;
    powerLimit: number;
    reactiveEnergy: boolean;
    reserveT1: string;
    reserveT2: string;
    reserveT3: null;
    reserveT4: null;
    status: boolean;
    t1: string;
    t2: string;
    t3: null;
    t4: null;
    tariffScheduleDate: string;
    totalEnergy: boolean;
    transitionDelay: number;
    voltageLimit: number;
  };
  description: {
    manufacturer: string;
    meterName: string;
    setUpOrganization: string;
    verificationDate: string;
  };
  information: {
    batteryVoltage: number;
    binCode: string;
    disbalance: boolean;
    meterConstant: number;
    programVersion: string;
    programming: boolean;
    relay: boolean;
    temperature: number;
    timeInNetwork: number;
  };
};
// ...
export type GPRSuserInfoValuesType = {
  city: string;
  district: string;
  email: string;
  fathersname: string;
  firstname: string;
  flat: string;
  house: string;
  lastname: string;
  organizationName: string;
  personalAccountNumber: string;
  phoneNumber: string;
  position: string;
  roleName: string;
  street: string;
};

export type GPRSListResponseType = {
  selected: boolean;
  roleName: string;
  fullName: string;
  personalAccountNumber: string;
  organizationName: string;
  location: string;
  address: string;
  active: boolean;
  activeEnergy: boolean;
  amperageLimit: number;
  apnName: string;
  apnPassword: string;
  apnUsername: string;
  batteryVoltage: number;
  binCode: string;
  closeDay: number;
  closeHour: number;
  createdAt: string;
  dateIndication: boolean;
  deviceId: number;
  disbalance: boolean;
  displayIndication: boolean;
  domainIp: string;
  id: string;
  impulseNumber: number;
  instantPower: boolean;
  ip: string;
  kwLimit: null;
  lastFixDate: string;
  limitDateFrom: null;
  limitDateTo: null;
  localIp: string;
  manufacturer: null;
  maxPower: boolean;
  meterConstant: number;
  meterId: string;
  meterName: null;
  meterNumberIndication: true;
  networkSettings: boolean;
  notified: boolean;
  pagingDelay: number;
  payloadLimit: number;
  periodLimitType: null;
  phaseAActivePowerIndication: boolean;
  phaseAAmperageIndication: boolean;
  phaseAPowerFactorIndication: boolean;
  phaseAVoltageIndication: boolean;
  phaseBActivePowerIndication: boolean;
  phaseBAmperageIndication: boolean;
  phaseBPowerFactorIndication: boolean;
  phaseBVoltageIndication: boolean;
  phaseCActivePowerIndication: boolean;
  phaseCAmperageIndication: boolean;
  phaseCPowerFactorIndication: boolean;
  phaseCVoltageIndication: boolean;
  phaseNAmperageIndication: boolean;
  pingIp: string;
  port: number;
  positiveActiveRateFlatPowerIndication: boolean;
  positiveActiveRatePeakEnergyIndication: boolean;
  positiveActiveRateTipEnergyIndication: boolean;
  positiveActiveRateValleyEnergyIndication: boolean;
  positiveActiveTotalEnergyIndication: boolean;
  positiveReactivePowerIndication: boolean;
  positiveReactivePowerPeakEnergyIndication: boolean;
  positiveReactivePowerRateFlatPowerIndication: boolean;
  positiveReactivePowerRateIndication: boolean;
  positiveReactivePowerRateValleyEnergyIndication: boolean;
  powerLimit: number;
  programVersion: string;
  programming: boolean;
  reactiveEnergy: boolean;
  relay: boolean;
  reserveT1: string;
  reserveT2: string;
  reserveT3: null;
  reserveT4: null;
  reverseActiveRateFlatPowerIndication: boolean;
  reverseActiveRatePeakEnergyIndication: boolean;
  reverseActiveRateTipEnergyIndication: boolean;
  reverseActiveRateValleyEnergyIndication: boolean;
  reverseActiveTotalEnergyIndication: boolean;
  reverseReactivePowerRateIndication: boolean;
  reverseReactiveRateFlatPowerIndication: boolean;
  reverseReactiveRatePeakEnergyIndication: boolean;
  reverseReactiveRateValleyEnergyIndication: boolean;
  reverseReactiveTotalEnergyIndication: boolean;
  setUpOrganization: null;
  status: boolean;
  t1: string;
  t2: string;
  t3: null;
  t4: null;
  tariffScheduleDate: string;
  temperature: number;
  timeInNetwork: number;
  timeIndication: boolean;
  totalActiveMaxDemandIndication: boolean;
  totalActivePowerIndication: boolean;
  totalEnergy: boolean;
  totalPowerFactorIndication: boolean;
  totalReactivePowerIndication: boolean;
  totalTimeInNetworkIndication: boolean;
  transitionDelay: number;
  updatedAt: string;
  userInfo: null | {
    personalAccountNumber: string;
    roleName: string;
    firstname: string;
    lastname: string;
    fathersname: string;
    organizationName: string;
    position: string;
    phoneNumber: string;
    email: string;
    city: string;
    district: string;
    street: string;
    house: string;
    flat: string;
  };
  verificationDate: null;
  versionIndication: false;
  voltageLimit: number;
};